+++ simapp/app_di.go
@@ -7,17 +7,22 @@ import (
 	"fmt"
 	"io"
 
+	_ "github.com/jackc/pgx/v5/stdlib" // Import and register pgx driver
+
 	clienthelpers "cosmossdk.io/client/v2/helpers"
 	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/core/registry"
 	corestore "cosmossdk.io/core/store"
 	"cosmossdk.io/depinject"
+	_ "cosmossdk.io/indexer/postgres" // register the postgres indexer
 	"cosmossdk.io/log"
 	"cosmossdk.io/x/accounts"
 	basedepinject "cosmossdk.io/x/accounts/defaults/base/depinject"
 	lockupdepinject "cosmossdk.io/x/accounts/defaults/lockup/depinject"
 	multisigdepinject "cosmossdk.io/x/accounts/defaults/multisig/depinject"
+	"cosmossdk.io/x/accounts/testing/account_abstraction"
+	"cosmossdk.io/x/accounts/testing/counter"
 	bankkeeper "cosmossdk.io/x/bank/keeper"
 	circuitkeeper "cosmossdk.io/x/circuit/keeper"
 	consensuskeeper "cosmossdk.io/x/consensus/keeper"
@@ -93,8 +98,10 @@ func init() {
 // AppConfig returns the default app config.
 func AppConfig() depinject.Config {
 	return depinject.Configs(
-		appConfig,                               // Alternatively use appconfig.LoadYAML(AppConfigYAML)
-		depinject.Provide(ProvideExampleMintFn), // optional: override the mint module's mint function with epoched minting
+		appConfig, // Alternatively use appconfig.LoadYAML(AppConfigYAML)
+		depinject.Provide(
+			ProvideExampleMintFn, // optional: override the mint module's mint function with epoched minting
+		),
 	)
 }
 
@@ -178,6 +185,10 @@ func NewSimApp(
 				//				return fmt.Errorf("invalid pub key size")
 				//			}
 				// 		})
+
+				// TESTING: do not add below account types
+				counter.ProvideAccount,
+				account_abstraction.ProvideAccount,
 			),
 		)
 	)
@@ -294,14 +305,15 @@ func (app *SimApp) setCustomAnteHandler() {
 	anteHandler, err := NewAnteHandler(
 		HandlerOptions{
 			ante.HandlerOptions{
-				AccountKeeper:      app.AuthKeeper,
-				BankKeeper:         app.BankKeeper,
-				ConsensusKeeper:    app.ConsensusParamsKeeper,
-				SignModeHandler:    app.txConfig.SignModeHandler(),
-				FeegrantKeeper:     app.FeeGrantKeeper,
-				SigGasConsumer:     ante.DefaultSigVerificationGasConsumer,
-				UnorderedTxManager: app.UnorderedTxManager,
-				Environment:        app.AuthKeeper.Environment,
+				AccountKeeper:            app.AuthKeeper,
+				BankKeeper:               app.BankKeeper,
+				ConsensusKeeper:          app.ConsensusParamsKeeper,
+				SignModeHandler:          app.txConfig.SignModeHandler(),
+				FeegrantKeeper:           app.FeeGrantKeeper,
+				SigGasConsumer:           ante.DefaultSigVerificationGasConsumer,
+				UnorderedTxManager:       app.UnorderedTxManager,
+				Environment:              app.AuthKeeper.Environment,
+				AccountAbstractionKeeper: app.AccountsKeeper,
 			},
 			&app.CircuitBreakerKeeper,
 		},