+++ client/v2/CHANGELOG.md
@@ -44,9 +44,11 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * [#18461](https://github.com/cosmos/cosmos-sdk/pull/18461) Support governance proposals.
 * [#20623](https://github.com/cosmos/cosmos-sdk/pull/20623) Introduce client/v2 tx factory.
 * [#20623](https://github.com/cosmos/cosmos-sdk/pull/20623) Extend client/v2 keyring interface with `KeyType` and `KeyInfo`.
+* [#22282](https://github.com/cosmos/cosmos-sdk/pull/22282) Added custom broadcast logic.
 
 ### Improvements
 
+* [#22443](https://github.com/cosmos/cosmos-sdk/pull/22443) Add keyring flags in query commands.
 * [#21936](https://github.com/cosmos/cosmos-sdk/pull/21936) Print possible enum values in error message after an invalid input was provided.
 
 ### API Breaking Changes